requirements.txt
Django==4.2.7
tensorflow==2.18.0
opencv-contrib-python==4.8.1.78
numpy==1.26.4
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.10.3
psycopg2-binary==2.9.9
cloud-tpu-client

# matplotlib 3.8.2 requires numpy<2,>=1.21, but you have numpy 2.0.2 which is incompatible.
# scikit-learn 1.3.2 requires numpy<2.0,>=1.17.3, but you have numpy 2.0.2 which is incompatible.